fix: resolve circular dependency by moving doingChat to stores.svelte.ts
