+++ src/lib/ChatScreens/DefaultChatScreen.svelte
@@ -8,7 +8,8 @@
     import { type Message } from "../../ts/storage/database.svelte";
     import { DBState } from 'src/ts/stores.svelte';
     import { getCharImage } from "../../ts/characters";
-    import { chatProcessStage, doingChat, sendChat } from "../../ts/process/index.svelte";
+    import { chatProcessStage, sendChat } from "../../ts/process/index.svelte";
+    import { doingChat } from "../../ts/stores.svelte";
     import { sleep } from "../../ts/util";
     import { language } from "../../lang";
     import { isExpTranslator, translate } from "../../ts/translator/translator";
