+++ src/lib/ChatScreens/Suggestion.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 	import { requestChatData } from "src/ts/process/request/request";
-    import { doingChat, type OpenAIChat } from "../../ts/process/index.svelte";
+    import { type OpenAIChat } from "../../ts/process/index.svelte";
+    import { doingChat } from "../../ts/stores.svelte";
     import { type character, type Message, type groupChat } from "../../ts/storage/database.svelte";
 	import { DBState } from 'src/ts/stores.svelte';
     import { selectedCharID } from "../../ts/stores.svelte";
