+++ src/ts/plugins/apiV3/v3.svelte.ts
@@ -17,7 +17,8 @@ import { getLLMCache, searchLLMCache } from "src/ts/translator/translator";
 import { hasher } from "src/ts/parser/parser.svelte";
 import localforage from "localforage";
 import { LLMFlags, LLMFormat, LLMProvider, LLMTokenizer, type LLMModel } from "src/ts/model/types";
-import { sendChat as processSendChat, doingChat } from "src/ts/process/index.svelte";
+import { sendChat as processSendChat } from "src/ts/process/index.svelte";
+import { doingChat } from "src/ts/stores.svelte";
 import { getModelInfo } from "src/ts/model/modellist";
 import type { ModelModeExtended } from "src/ts/process/request/shared";
 import { requestChatDataMain } from "src/ts/process/request/request";
